requirements.txt
pandas
pydantic
langchain-chroma
langchain-community
langchain-ollama
langchain-chroma
streamlit
langchain-community
langchain-openai